Group conditionals in Applicator.
It was reported in #5 that singularity was erroring out when running
apply due to the action variable not being an iterable (which it's not
if it's not set).  This should remove that error and fix the problem.

If singularity is popular enough I need to begin work on a version 2.x
that removes the complications provided by this version but still uses
the modular nature.  In fact, it should probably be a custom reader and
action set for cloud-init in all honesty.